fix: app перенести функции в отдельный файл
